pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-rate"
version = "0.1.0"
description = "A machine learning based face rating website"
authors = [
    {name = "Daron Lebaredian", email = "daron.lebaredian@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "torch==2.2.2",
    "torchvision==0.17.2",
    "facenet-pytorch==2.6.0",
    "Pillow==10.2.0",
    "python-multipart==0.0.12",
    "httpx==0.28.0",
    "numpy==1.26.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "black==24.8.0",
    "flake8==7.1.1",
    "python-dotenv>=1.0.1",
]

[project.urls]
Homepage = "https://github.com/Prototype19/Face-Rate"
Repository = "https://github.com/Prototype19/Face-Rate"
Issues = "https://github.com/Prototype19/Face-Rate/issues"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
